Allow larger length seed entropy

+++ tournament.py
@@ -42,7 +42,7 @@ def get_player_lists(family_size, repeat_count):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--result_dir", default="results", help="Directory path to dump results")
-    parser.add_argument("--seed_entropy", "-s", type=np.uint64, help="Seed used to generate seed for each game")
+    parser.add_argument("--seed_entropy", "-s", type=int, help="Seed used to generate seed for each game")
     parser.add_argument("--family_size", "-f", default=[2, 3, 4, 6, 8, 9, 12], nargs="+", type=int, help="List of size of family")
     args = parser.parse_args()
     RESULT_DIR = args.result_dir
